% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvMixedIC50.R
\name{fitModel}
\alias{fitModel}
\title{Curve fitting the data to the model}
\usage{
fitModel(gDat, vStart = c(8.886464, 1.495953), bLargeScale = TRUE,
  bSilent = TRUE)
}
\arguments{
\item{gDat}{A \code{\link{groupedData}} object prepared in \code{\link{initIC50}}}

\item{vStart}{The starting values of for the curve fit.}

\item{bLargeScale}{Whether the covariance between the position and scale parameter on the cell line level are assumed to be correlated or not. In a large scale screen this is set to true as it further stabilizes the fit. In small bespoke screens this is set to false as the model otherwise struggles to converge.}

\item{bSilent}{If set to true (default), suppresses status output.}
}
\value{
fitted model, a nlme object
}
\description{
Curve fitting the data prepared with \code{\link{initIC50}} to the model. The resultant object is further processed by \code{\link{gatherModelStats}}.
}
\examples{
gDat      <- initIC50(szFileName='~/myFile.csv')
fmMod1    <- fitModel(gDat)
outStats  <- gatherModelStats(gDat=gDat,fmMod1=fmMod1)
dfIC50    <- outStats$IC50
dfRMSE    <- outStats$RMSE
dfAUC     <- outStats$AUC
dfAUCtrap <- outStats$AUCtrap

}

