% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvMixedIC50.R
\name{gatherModelStats}
\alias{gatherModelStats}
\title{Gathering model statistics}
\usage{
gatherModelStats(gDat, fmMod1)
}
\arguments{
\item{gDat}{A \code{\link{groupedData}} object prepared in \code{\link{initIC50}}}

\item{fmMod1}{The model fitted with \code{\link{fitModel}}}
}
\value{
A list with named items, holding the IC50s, AUCs, AUCtrap, RMSE, the gDat, the mcv, and the fitted model.
}
\description{
The data prepared with \code{\link{initIC50}}, and fitted with \code{\link{fitModel}} to the model, is used here to assemble output statiscs. As \code{\link{fitModel}} only parametrizes the model, the relevant statistics are here extracted. These values include the IC50, the RMSE, the model based AUC, and the trapezoid AUC.
When the series contains replicates, the normal AUC is define, but the trapezoid AUC is not.
}
\examples{
gDat      <- initIC50(szFileName='~/myFile.csv')
fmMod1    <- fitModel(gDat)
outStats  <- gatherModelStats(gDat=gDat,fmMod1=fmMod1)
dfIC50    <- outStats$IC50
dfRMSE    <- outStats$RMSE
dfAUC     <- outStats$AUC
dfAUCtrap <- outStats$AUCtrap

}

