% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvMixedIC50.R
\name{getXfromConcSeries}
\alias{getXfromConcSeries}
\title{Convenience function for dilution series coding}
\usage{
getXfromConcSeries(fDilutionSeries)
}
\arguments{
\item{fDilutionSeries}{A series of decreasing fractions.}
}
\value{
Series of predictor values
}
\description{
A convenience function for getting the predictor values for custom dilution series. The function takes as argument a decreasing vector of fractions and converts them to the internal coding scheme of the curve fit, which defines 9 to be the highest screening concentration. The default assumption is a two fold dilution, and this function is there to accommodate other divisors. The series of 1, 1/2, 1/4, 1/8 would result in 9, 8, 7, 6.
}
\examples{
\% example of how to get x for a sqrt(2)-fold dilution per step
xSeries <- getXfromConcSeries(c(1, 0.7071, 0.5, 0.35355, 0.25))

}

