% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/djvMixedIC50.R
\name{plotSingleResponse}
\alias{plotSingleResponse}
\title{Plotting a single dose response curve}
\usage{
plotSingleResponse(gDat, mc1, szCL, szDrug, dfOld = NULL, szPrefix = "",
  szDrugName = "")
}
\arguments{
\item{gDat}{A groupedData object prepared in \code{\link{initIC50}}}

\item{mc1}{The internal coefficient of the nonlinear model, obtainable by \code{mc1 <- coef(fmMod1)}, using \code{\link{coef}} from \code{\link{nlme}}}

\item{szCL}{The name/identifier of the cell line}

\item{szDrug}{The name/identifier of the drug}

\item{dfOld}{Optional argument for supplying reference values}

\item{szPrefix}{Optional parameter for prefixing the plot title}

\item{szDrugName}{Optional parameter for the drug name, printed as part of plot title}
}
\value{
No values are returned
}
\description{
The data prepared with \code{\link{initIC50}}, and fitted with \code{\link{fitModel}} to the model, is used here to assemble output statistics. As \code{\link{fitModel}} only parametrizes the model, the relevant statistics are here extracted. These values include the IC50, the RMSE, the model based AUC, and the trapezoid AUC.
}
\examples{
plotSingleResponse(gDat,mc1,'MC-CAR', 'DrugID1005',szPrefix='Screen 1',szDrugName='Platinum')

}

